xarray.backends.plugins.get_backend: fix passing in a backend directly
The `open_dataset` and `open_mfdataset` functions allow a user to pass
in a (custom) backend directly. This was however not functioning.

Selecting an engine with a string was functioning as expected, however,
passing in a backend class directly failed.

An instance of the backend is now created. A test has been added, as
well as some type hints.